Create a `Compornent/Icons` folder. Moved 6 kinds of self-made icons to new folders. Corrected `@import` path where the corresponding icon is used.
