+++ src/client/js/components/Icons/DarkModeIcon.jsx
@@ -9,12 +9,10 @@ const DarkModeIcon = () => (
   >
     <g transform="translate(-923.5 -688.5)">
       <rect width="23" height="23" fill="none" transform="translate(923.5 688.5)" />
-      {/* <g className="darkmode-icon"> */}
       <path d="M934.893,710.532a10.646,10.646,0,0,1-10.378-8.416.7.7,0,0,1,1.138-.686,
        7.621,7.621,0,0,0,10.721-10.744.7.7,0,0,1,.683-1.14,10.6,10.6,0,0,1-2.164,
         20.986Zm-8.417-6.9A9.2,9.2,0,1,0,938.583,691.5a9.028,9.028,0,0,1-12.107,12.133Z"
       />
-      {/* </g> */}
     </g>
   </svg>
