+++ src/client/js/components/Icons/EditNewPageIcon.jsx
