+++ src/client/js/components/Icons/LightModeIcon.jsx
