+++ src/client/js/components/Icons/ReturnTopIcon.jsx
