+++ src/client/js/components/Icons/SidebarDockIcon.jsx
