+++ src/client/js/components/Icons/SidebarDrawerIcon.jsx
